+++ R/read_agilent_dx.R
@@ -18,6 +18,9 @@
 #' @author Ethan Bass
 #' @return A chromatogram in the format specified by `format_out` (retention
 #' time x wavelength).
+#' @examples \dontrun{
+#' read_agilent_dx(path)
+#' }
 #' @author Ethan Bass
 #' @family 'Agilent' parsers
 #' @export
@@ -73,8 +76,8 @@ read_agilent_dx <- function (path,  what = c("chroms", "dad"), path_out = NULL,
   }
   if (any(what == "instrument")){
     if (length(files.path$instrument) > 0){
-      instrument <- lapply(files.path$instrument, read_chemstation_it, 
-                           format_out = format_out, data_format = data_format, 
+      instrument <- lapply(files.path$instrument, read_chemstation_it,
+                           format_out = format_out, data_format = data_format,
                            read_metadata = read_metadata,
                            metadata_format = metadata_format, source_file = path)
       names(instrument) <- sapply(instrument, function(x) attr(x, "detector_range"))